+++ src/third/DvaComponent.ts
@@ -1,0 +1,3 @@
+import { SimpleThirdComponent } from "./SimpleThirdComponent";
+
+export class DvaComponent extends SimpleThirdComponent { }