+++ src/third/QaComponent.ts
@@ -1,0 +1,3 @@
+import { SimpleThirdComponent } from "./SimpleThirdComponent";
+
+export class QaComponent extends SimpleThirdComponent { }